+++ src/main/java/buildcraft/robotics/boards/BoardRobotFluidCarrierNBT.java
@@ -45,11 +45,6 @@ public class BoardRobotFluidCarrierNBT extends RedstoneBoardRobotNBT {
     }
 
     @Override
-    public String getItemModelLocation() {
-        return ID;
-    }
-
-    @Override
     public RedstoneBoardRobot create(IRobotAccess robot) {
         return new BoardRobotFluidCarrier(robot);
     }